Remove volumes from compose
Docker volumes, unfortunately, are somewhat useless as implemented.
Mounting them nukes the contents in the container, making the
container dependent on the host, which is just stupid and makes the
container a PITA to use. The workaround is to use a startup script
to populate the volumes in the container after the container starts,
but that doesn't play well with supervisor. For now, just drop the
volumes from docker-compose.yaml so the container works, but the
host can't access the data in the volumes.